feat(catalog): move repository adapters to domain bounded context
