--- core/src/adapters/outbound/database/catalog/MongoCatalogUnitOfWorkAdapter.ts
@@ -1,4 +1,0 @@
-import { CatalogUnitOfWorkPort } from '../../../../domains/catalog';
-import { MongoUnitOfWorkBase } from '../MongoUnitOfWorkBase';
-
-export class MongoCatalogUnitOfWorkAdapter extends MongoUnitOfWorkBase implements CatalogUnitOfWorkPort {}
