+++ core/src/domains/catalog/adapters/outbound/database/MongoCatalogUnitOfWorkAdapter.ts
@@ -1,0 +1,4 @@
+import { CatalogUnitOfWorkPort } from '../../..';
+import { MongoUnitOfWorkBase } from '../../../../../adapters/outbound/database/MongoUnitOfWorkBase';
+
+export class MongoCatalogUnitOfWorkAdapter extends MongoUnitOfWorkBase implements CatalogUnitOfWorkPort {}
